Made it slightly more realistic; added time line; minor stuff.
I modified the project plan a bit, to make it slightly less over-ambitious.
In particular I removed the radiation hydrodynamics part, which, I think,
is indeed too much and may risk to absorb all the time of the postdoc, and
remove that time from the rest of the project.

I also now implemented the time line.

I added some todos regarding which part of the finances should be Z-only.

I changed to pdflatex, so now all figures are converted to PDF.