I added some text to D2 on infall-disk interaction... Difficult topic.
I am contemplating to remove the shadow-induced spiral stuff and instead
concentrate on the infall-disk interaction. The shadow-induced spiral
stuff is already published by Montesinos et al. Maybe, just maybe, we
can add some new features to this, but I consider it unlikely that this
issue will still be current by the time the postdoc will come to it.
Instead, the infall-disk interaction is something completely new, albeit
of course studied in the big star formation simulations.