D2: Removed shadow-induced spiral subproject, added: infall-disk interaction.
I made some big modification to D2. The shadow-induced spiral project is,
I think, less interesting because in 2 years this question will have been
solved. Instead, I think it is more important to study the interaction
between the secondary infall with an already pre-existing disk at a very
different angular momentum direction. I reorganized a bit the D2 project
and the questions to be addressed, and removed the shadow-induced part
and added the interaction part as a separate sub-project.